+++ drizzle/0007_sad_morlocks.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `storage_locations` ADD `active` integer DEFAULT true NOT NULL;